fixup: splitting up the alist files
